requirements.txt
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
Django~=4.2.3
django-environ==0.10.0
pytest~=7.4.0
django-dynamic-fixture==3.1.2
pytest-cov==4.0.0
pytest-django==4.5.2
requests==2.28.1
requests-mock==1.10.0